Only update a cookie value, not attributes when using page.cookies
It appears that if a cookie on the page has an expiry other session cookies on
the page (cookies without an expiry value) will be incorrectly assigned that
expiry when attempting to retrieve the cookies via Phantom's `page.cookie`.

As a compromise we'll only update a cookie's value in case a script on the page
modified the original value specified by the request headers.